+++ storefront/src/modules/layout/components/cart-dropdown/index.tsx
@@ -136,6 +136,7 @@ const CartDropdown = ({
                             thumbnail={item.thumbnail}
                             images={item.variant?.product?.images}
                             size="square"
+                            alt={item.title || item.product_title || "Product photo"}
                           />
                         </LocalizedClientLink>
                         <div className="flex flex-col justify-between flex-1">