+++ storefront/src/modules/store/templates/index.tsx
@@ -19,7 +19,7 @@ const StoreTemplate = ({
   optionValueIds?: OptionValueIds
 }) => {
   const pageNumber = page ? parseInt(page) : 1
-  const sort = sortBy || "created_at"
+  const sort = sortBy || "featured"
 
   return (
     <div